Added a global required organization option
In our case here at @ContaAzul, we run our own mention-bot to support
our private repos.

In that case, it would be great to set a required organization globally
instead of configuring it for each repo (we have lot's of them).

This PR should fix that, don't know if it's the best way of doing this,
though.