Modify installation instructions
I'm brand-new to mention-bot so please forgive me if I am incorrect. :]

Following the defualt instructions did not work for me. I couldn't find
any instance in which SSL support on the mention-bot end is supported.
I had to disable it in the GitHub hook for mention-bot to work.

I also noted that the default URL for the GitHub webhook needs to be
TCP/5000.