Clarify default message (#233)
The previous default message always stated that the history of files in the pull request was analyzed. This is not the case when findPotentialReviewers is set to false - reviewers get mentioned based on settings in the configuration file. This change makes the message more accurate when the setting is disabled.